feat: implement `Portfolio._perform_entry()` method.
This method is the entry action performed when processing a trade.
The opposite action is `Portfolio._perform_exit()` that do the reversed operation.
I finally ended up creating `WealthSeries` to manage the wealth of a portfolio.
This creates a lot of code; it looks like cleared when the logic is not gathered in a single place, though.